bigfix

+++ vm/conf/conf.hpp
@@ -12,10 +12,10 @@ namespace LLCCEP_vm {
 	struct config {
 		::std::vector<::std::string> dev;
 
-		uint16_t displayW;
-		uint16_t displayH;
+		int displayW;
+		int displayH;
 
-		size_t ramS;
+		long long unsigned ramS;
 	};
 
 	config read_configuration_file(::std::string path);
